reconstitution siamese
