+++ .unotes/unotes_meta.json
@@ -1,0 +1,45 @@
+{
+  "name": "",
+  "isOrdered": true,
+  "folders": {
+    "__pycache__": {
+      "name": "__pycache__",
+      "isOrdered": true,
+      "folders": {},
+      "files": {}
+    },
+    "assets": {
+      "name": "assets",
+      "isOrdered": true,
+      "folders": {},
+      "files": {}
+    },
+    "data": {
+      "name": "data",
+      "isOrdered": true,
+      "folders": {},
+      "files": {}
+    },
+    "model": {
+      "name": "model",
+      "isOrdered": true,
+      "folders": {},
+      "files": {}
+    },
+    "Summaries": {
+      "name": "Summaries",
+      "isOrdered": true,
+      "folders": {},
+      "files": {}
+    },
+    "weights": {
+      "name": "weights",
+      "isOrdered": true,
+      "folders": {},
+      "files": {}
+    }
+  },
+  "files": {
+    "README": 0
+  }
+}
